+++ packages/core/ui/src/index.ts
@@ -4,6 +4,8 @@ export * from '@srcube-taro/input'
 
 export * from '@srcube-taro/layouts'
 
+export * from '@srcube-taro/modal'
+
 export * from '@srcube-taro/spinner'
 
 export * from '@srcube-taro/theme'