+++ packages/ui/modal/src/index.ts
@@ -1,0 +1,15 @@
+export type { ModalProps } from './modal'
+export { default as Modal } from './modal'
+
+export type { ModalBodyProps } from './modal-body'
+export { default as ModalBody } from './modal-body'
+export type { ModalContentProps } from './modal-content'
+export { default as ModalContent } from './modal-content'
+export type { ModalFooterProps } from './modal-footer'
+export { default as ModalFooter } from './modal-footer'
+export type { ModalHeaderProps } from './modal-header'
+export { default as ModalHeader } from './modal-header'
+
+// hooks
+export { useDisclosure } from '@srcube-taro/hooks'
+export { usePageScrollLock } from '@srcube-taro/hooks'