+++ packages/ui/button/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @srcube-taro/button
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @srcube-taro/theme@0.0.4
+  - @srcube-taro/spinner@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes