+++ packages/ui/input/CHANGELOG.md
@@ -1,10 +1,17 @@
 # @srcube-taro/input
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @srcube-taro/theme@0.0.4
+
 ## 0.0.1
 
 ### Patch Changes
 
-- Components:
-  - input
+- Components
+  input
 - Updated dependencies
   - @srcube-taro/theme@0.0.3