+++ packages/ui/layouts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @srcube-taro/layouts
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @srcube-taro/theme@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes